wifi

+++ main/control/rotation_matrix.h
@@ -7,7 +7,7 @@
 
 class RotationMatrix {
 
-    int8_t x_gravity_factor, y_gravity_factor, z_gravity_factor;
+    float x_gravity_factor, y_gravity_factor, z_gravity_factor;
     MsTimer timer {};
     // rotation of the drone in world coordinates
     float matrix[9] = {1, 0, 0, 0, 1, 0, 0, 0, 1};
